% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{adjusted_R2}
\alias{adjusted_R2}
\title{Compute adjusted R-squared}
\usage{
adjusted_R2(R2, n, p)
}
\arguments{
\item{R2}{R-squared value}

\item{n}{Sample size}

\item{p}{Number of predictors}
}
\value{
Adjusted R-squared
}
\description{
Compute adjusted R-squared
}
\keyword{internal}
