% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{checkHTML}
\alias{checkHTML}
\title{Check HTML files for statistical consistency (DEFUNCT)}
\usage{
checkHTML(files, messages = TRUE, ...)
}
\arguments{
\item{files}{Defunct argument.}

\item{messages}{Defunct argument.}

\item{...}{Defunct argument.}
}
\value{
Errors with a migration message.
}
\description{
Removed in effectcheck 0.4.0. HTML can be passed directly to \code{\link[=check_text]{check_text()}}
via \code{check_text(rvest::html_text2(xml2::read_html(path)))}.
}
\keyword{internal}
