% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{checkHTMLdir}
\alias{checkHTMLdir}
\title{Check a directory of HTML files (DEFUNCT)}
\usage{
checkHTMLdir(dir, subdir = TRUE, messages = TRUE, ...)
}
\arguments{
\item{dir}{Defunct argument.}

\item{subdir}{Defunct argument.}

\item{messages}{Defunct argument.}

\item{...}{Defunct argument.}
}
\value{
Errors with a migration message.
}
\description{
Removed in effectcheck 0.4.0.
}
\keyword{internal}
