% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{checkPDFdir}
\alias{checkPDFdir}
\title{Check a directory of PDF files (DEFUNCT)}
\usage{
checkPDFdir(
  dir,
  subdir = TRUE,
  try_tables = TRUE,
  try_ocr = FALSE,
  messages = TRUE,
  ...
)
}
\arguments{
\item{dir}{Defunct argument.}

\item{subdir}{Defunct argument.}

\item{try_tables}{Defunct argument.}

\item{try_ocr}{Defunct argument.}

\item{messages}{Defunct argument.}

\item{...}{Defunct argument.}
}
\value{
Errors with a migration message.
}
\description{
Removed in effectcheck 0.4.0.
}
\keyword{internal}
