% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{ci_adjusted_R2}
\alias{ci_adjusted_R2}
\title{CI for adjusted R-squared (monotone transform of the R^2 CI)}
\usage{
ci_adjusted_R2(R2, N, k, df1, df2, level = 0.95)
}
\description{
CI for adjusted R-squared (monotone transform of the R^2 CI)
}
\keyword{internal}
