% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{ci_cohens_f}
\alias{ci_cohens_f}
\title{Calculate CI for Cohen's f}
\usage{
ci_cohens_f(F_val, df1, df2, level = 0.95)
}
\arguments{
\item{F_val}{F statistic}

\item{df1}{df1}

\item{df2}{df2}

\item{level}{CI level}
}
\value{
ci_result list
}
\description{
Derived from Partial Eta-Squared CI.
}
\keyword{internal}
