% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{ci_d_ind_all}
\alias{ci_d_ind_all}
\title{Compute CIs for independent d via all available methods}
\usage{
ci_d_ind_all(d, n1, n2, level = 0.95)
}
\arguments{
\item{d}{Cohen's d}

\item{n1}{Sample size 1}

\item{n2}{Sample size 2}

\item{level}{Confidence level}
}
\value{
Named list of ci_result objects (one per successful method)
}
\description{
Compute CIs for independent d via all available methods
}
\keyword{internal}
