% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{ci_d_ind_approx}
\alias{ci_d_ind_approx}
\title{Approximate CI for independent d}
\usage{
ci_d_ind_approx(d, n1, n2, level = 0.95)
}
\arguments{
\item{d}{Cohen's d}

\item{n1}{Sample size 1}

\item{n2}{Sample size 2}

\item{level}{Confidence level (default 0.95)}
}
\value{
Vector of bounds \link{lower, upper}
}
\description{
Uses Hedges/CMC large-sample approximation.
}
\keyword{internal}
