% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{ci_dz_all}
\alias{ci_dz_all}
\title{Compute CIs for paired dz via all available methods}
\usage{
ci_dz_all(dz, n, level = 0.95)
}
\arguments{
\item{dz}{Cohen's dz}

\item{n}{Sample size (number of pairs)}

\item{level}{Confidence level}
}
\value{
Named list of ci_result objects
}
\description{
Compute CIs for paired dz via all available methods
}
\keyword{internal}
