% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{ci_etap2}
\alias{ci_etap2}
\title{Calculate CI for Partial Eta-Squared}
\usage{
ci_etap2(F_val, df1, df2, level = 0.95)
}
\arguments{
\item{F_val}{F statistic}

\item{df1}{df1}

\item{df2}{df2}

\item{level}{CI level}
}
\value{
ci_result list
}
\description{
Calculate CI for Partial Eta-Squared
}
\keyword{internal}
