% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectcheck-class.R
\name{compare_to_variants}
\alias{compare_to_variants}
\title{Compare reported value to all variants}
\usage{
compare_to_variants(x, row_index = 1)
}
\arguments{
\item{x}{An effectcheck object}

\item{row_index}{The row index}
}
\value{
A data frame with variant comparisons
}
\description{
Creates a comparison table showing the reported value against all computed variants.
}
\examples{
res <- check_text("t(28) = 2.21, p = .035, d = 0.80")
compare_to_variants(res, 1)
}
