% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_dav_range}
\alias{compute_dav_range}
\title{Compute range of plausible dav values}
\usage{
compute_dav_range(dz, r_grid = c(seq(0.1, 0.9, by = 0.1), 0.95))
}
\arguments{
\item{dz}{Cohen's dz}

\item{r_grid}{Vector of correlations to test}
}
\value{
List with min, max, median, and values
}
\description{
Calculates dav across a grid of possible correlations.
}
\keyword{internal}
