% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{d_ind_from_t}
\alias{d_ind_from_t}
\title{Calculate Cohen's d from t-statistic (Independent Samples)}
\usage{
d_ind_from_t(t, n1, n2)
}
\arguments{
\item{t}{t-statistic}

\item{n1}{Sample size 1}

\item{n2}{Sample size 2}
}
\value{
Cohen's d
}
\description{
Calculate Cohen's d from t-statistic (Independent Samples)
}
\keyword{internal}
