% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{dav_from_dz}
\alias{dav_from_dz}
\title{Convert dz to dav (Cohen's d for average variance)}
\usage{
dav_from_dz(dz, r)
}
\arguments{
\item{dz}{Cohen's dz}

\item{r}{Correlation between measures}
}
\value{
Cohen's dav
}
\description{
Convert dz to dav (Cohen's d for average variance)
}
\keyword{internal}
