% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_TOL_EFFECT}
\alias{DEFAULT_TOL_EFFECT}
\title{Constants for EffectCheck}
\format{
An object of class \code{list} of length 28.
}
\usage{
DEFAULT_TOL_EFFECT
}
\description{
Default values and tolerances used across the package.
}
\keyword{internal}
