% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{drm_from_dz}
\alias{drm_from_dz}
\title{Convert dz to drm (Cohen's d for raw means)}
\usage{
drm_from_dz(dz, r = NA_real_)
}
\arguments{
\item{dz}{Cohen's dz}

\item{r}{Correlation (unused, for interface compatibility)}
}
\value{
Cohen's drm
}
\description{
Convert dz to drm (Cohen's d for raw means)
}
\keyword{internal}
