% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{dz_from_t}
\alias{dz_from_t}
\title{Calculate Cohen's dz from t-statistic (Paired)}
\usage{
dz_from_t(t, n)
}
\arguments{
\item{t}{t-statistic}

\item{n}{Sample size (number of pairs)}
}
\value{
Cohen's dz
}
\description{
Calculate Cohen's dz from t-statistic (Paired)
}
\keyword{internal}
