% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\docType{data}
\name{EFFECT_SIZE_FAMILIES}
\alias{EFFECT_SIZE_FAMILIES}
\title{Effect Size Type Definitions}
\format{
An object of class \code{list} of length 16.
}
\usage{
EFFECT_SIZE_FAMILIES
}
\description{
Maps reported effect size names to their family and variants.
Based on Guide to Effect Sizes and Confidence Intervals (Jane et al., 2024)
https://matthewbjane.quarto.pub/
}
\keyword{internal}
