% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{filter_by_delta}
\alias{filter_by_delta}
\title{Filter results by effect size delta}
\usage{
filter_by_delta(x, min_delta = 0, max_delta = Inf)
}
\arguments{
\item{x}{An effectcheck object}

\item{min_delta}{Minimum absolute delta to include (default 0)}

\item{max_delta}{Maximum absolute delta to include (default Inf)}
}
\value{
An effectcheck object containing only results within the delta range
}
\description{
Filters effectcheck results by the magnitude of effect size discrepancy.
}
\examples{
results <- check_text("t(28) = 2.21, p = .035, d = 0.80")
filter_by_delta(results, min_delta = 0.1)
}
