% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{fisher_ci_r}
\alias{fisher_ci_r}
\title{Fisher's z-transformation CI for correlation}
\usage{
fisher_ci_r(r, n, level = 0.95)
}
\arguments{
\item{r}{Correlation coefficient}

\item{n}{Sample size}

\item{level}{Confidence level}
}
\value{
Vector of bounds \link{lower, upper}
}
\description{
Fisher's z-transformation CI for correlation
}
\keyword{internal}
