% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{g_ind_from_t}
\alias{g_ind_from_t}
\title{Calculate Hedges' g from t-statistic}
\usage{
g_ind_from_t(t, n1, n2)
}
\arguments{
\item{t}{t-statistic}

\item{n1}{Sample size 1}

\item{n2}{Sample size 2}
}
\value{
Hedges' g
}
\description{
Calculate Hedges' g from t-statistic
}
\keyword{internal}
