% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectcheck-class.R
\name{get_alternatives}
\alias{get_alternatives}
\title{Get alternative suggestions for a row}
\usage{
get_alternatives(x, row_index = 1)
}
\arguments{
\item{x}{An effectcheck object}

\item{row_index}{The row index}
}
\value{
A list of alternative effect size suggestions
}
\description{
Get alternative suggestions for a row
}
\examples{
res <- check_text("t(28) = 2.21, p = .035, d = 0.80")
get_alternatives(res, 1)
}
