% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_decision_errors}
\alias{get_decision_errors}
\title{Get decision errors from effectcheck results}
\usage{
get_decision_errors(x)
}
\arguments{
\item{x}{An effectcheck object}
}
\value{
An effectcheck object containing only decision errors
}
\description{
Extracts results where the significance decision would be reversed
(i.e., reported as significant when computed is not, or vice versa).
}
\examples{
\dontrun{
results <- check_dir("manuscripts/")
decision_errors <- get_decision_errors(results)
}
}
