% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_errors}
\alias{get_errors}
\title{Get errors from effectcheck results}
\usage{
get_errors(x)
}
\arguments{
\item{x}{An effectcheck object}
}
\value{
An effectcheck object containing only errors
}
\description{
Convenience function to extract only ERROR status results.
}
\examples{
\dontrun{
results <- check_dir("manuscripts/")
errors <- get_errors(results)
}
}
