% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{get_ncp_F}
\alias{get_ncp_F}
\title{Find Non-Centrality Parameter (NCP) Confidence Limits for F-distribution}
\usage{
get_ncp_F(F_val, df1, df2, level = 0.95)
}
\arguments{
\item{F_val}{Observed F statistic}

\item{df1}{Numerator degrees of freedom}

\item{df2}{Denominator degrees of freedom}

\item{level}{Confidence level (default 0.95)}
}
\value{
Vector c(lambda_low, lambda_high)
}
\description{
Uses uniroot to invert the non-central F CDF.
}
\keyword{internal}
