% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectcheck-class.R
\name{get_same_type_variants}
\alias{get_same_type_variants}
\title{Get same-type variants for a row}
\usage{
get_same_type_variants(x, row_index = 1)
}
\arguments{
\item{x}{An effectcheck object}

\item{row_index}{The row index}
}
\value{
A list of same-type variants with their values and metadata
}
\description{
Get same-type variants for a row
}
\examples{
res <- check_text("t(28) = 2.21, p = .035, d = 0.80")
get_same_type_variants(res, 1)
}
