% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_tolerance}
\alias{get_tolerance}
\title{Get Tolerance Config}
\usage{
get_tolerance(type = c("effect", "ci", "p"))
}
\arguments{
\item{type}{Type of tolerance ("effect", "ci", "p")}
}
\description{
Helper to get tolerances, falling back to constants.
}
