% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_warnings}
\alias{get_warnings}
\title{Get warnings from effectcheck results}
\usage{
get_warnings(x)
}
\arguments{
\item{x}{An effectcheck object}
}
\value{
An effectcheck object containing only warnings
}
\description{
Convenience function to extract only WARN status results.
}
\examples{
\dontrun{
results <- check_dir("manuscripts/")
warnings <- get_warnings(results)
}
}
