% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{hedges_J}
\alias{hedges_J}
\title{Calculate Hedges' J correction factor}
\usage{
hedges_J(df)
}
\arguments{
\item{df}{Degrees of freedom}
}
\value{
Correction factor
}
\description{
Calculate Hedges' J correction factor
}
\keyword{internal}
