% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_error}
\alias{log_error}
\title{Log Error Message}
\usage{
log_error(msg, ...)
}
\arguments{
\item{msg}{Message string}

\item{...}{variables for interpolation}
}
\value{
Invisible NULL. Called for its side effect of logging.
}
\description{
Log Error Message
}
\keyword{internal}
