% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{numify}
\alias{numify}
\title{Convert string to numeric with warning suppression}
\usage{
numify(x)
}
\arguments{
\item{x}{String or vector}
}
\value{
Numeric value(s)
}
\description{
Convert string to numeric with warning suppression
}
\keyword{internal}
