% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{numify_int}
\alias{numify_int}
\title{Convert string to integer, stripping thousands-separator commas}
\usage{
numify_int(x)
}
\arguments{
\item{x}{String or vector}
}
\value{
Integer value(s)
}
\description{
Used ONLY for sample size values (N, n1, n2) where commas are always
thousands separators, never decimal commas.
}
\keyword{internal}
