% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{phi_from_chisq}
\alias{phi_from_chisq}
\title{Calculate phi coefficient from Chi-square}
\usage{
phi_from_chisq(chisq, N)
}
\arguments{
\item{chisq}{Chi-square statistic}

\item{N}{Total sample size}
}
\value{
Phi coefficient
}
\description{
Calculate phi coefficient from Chi-square
}
\keyword{internal}
