% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.effectcheck_comparison}
\alias{print.effectcheck_comparison}
\title{Print method for effectcheck comparison}
\usage{
\method{print}{effectcheck_comparison}(x, ...)
}
\arguments{
\item{x}{An effectcheck_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for effectcheck comparison
}
