% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectcheck-class.R
\name{print.summary.effectcheck}
\alias{print.summary.effectcheck}
\title{Print method for summary.effectcheck objects}
\usage{
\method{print}{summary.effectcheck}(x, ...)
}
\arguments{
\item{x}{A summary.effectcheck object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for summary.effectcheck objects
}
