% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{process_files_internal}
\alias{process_files_internal}
\title{Internal helper to process a list of files}
\usage{
process_files_internal(
  files,
  read_args = list(),
  check_args = list(),
  messages = TRUE
)
}
\arguments{
\item{files}{Character vector of file paths to process}

\item{read_args}{List of arguments for read_any_text()}

\item{check_args}{List of arguments for check_text()}

\item{messages}{Logical, show progress messages (default TRUE)}
}
\value{
An effectcheck object with results from all files
}
\description{
Internal helper to process a list of files
}
\keyword{internal}
