% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{r_from_t}
\alias{r_from_t}
\title{Calculate correlation r from t-statistic}
\usage{
r_from_t(t, df)
}
\arguments{
\item{t}{t-statistic}

\item{df}{Degrees of freedom}
}
\value{
Correlation r
}
\description{
Calculate correlation r from t-statistic
}
\keyword{internal}
