% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectcheck-class.R
\name{[.effectcheck}
\alias{[.effectcheck}
\title{Subset method for effectcheck objects}
\usage{
\method{[}{effectcheck}(x, ...)
}
\arguments{
\item{x}{An effectcheck object}

\item{...}{Subsetting arguments}
}
\value{
An effectcheck object
}
\description{
Preserves effectcheck class when subsetting.
}
\examples{
res <- check_text("t(28) = 2.21, p = .035. F(1, 50) = 4.03, p = .049")
res[1, ]
}
