% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{V_from_chisq}
\alias{V_from_chisq}
\title{Calculate Cramer's V from Chi-square}
\usage{
V_from_chisq(chisq, N, m)
}
\arguments{
\item{chisq}{Chi-square statistic}

\item{N}{Total sample size}

\item{m}{Smaller dimension - 1 (min(r-1, c-1))}
}
\value{
Cramer's V
}
\description{
Calculate Cramer's V from Chi-square
}
\keyword{internal}
