% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\docType{data}
\name{VARIANT_METADATA}
\alias{VARIANT_METADATA}
\title{Variant Metadata}
\format{
An object of class \code{list} of length 21.
}
\usage{
VARIANT_METADATA
}
\description{
Provides assumptions and usage information for each effect size variant.
}
\keyword{internal}
