% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{ci_eta2}
\alias{ci_eta2}
\title{Alias for ci_etap2 (used for eta2 when design implies equivalence or as approximation)}
\usage{
ci_eta2(F_val, df1, df2, level = 0.95)
}
\description{
Alias for ci_etap2 (used for eta2 when design implies equivalence or as approximation)
}
