% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{filter_by_source}
\alias{filter_by_source}
\title{Filter results by source file}
\usage{
filter_by_source(x, files, pattern = FALSE)
}
\arguments{
\item{x}{An effectcheck object}

\item{files}{Character vector of file names or patterns to include}

\item{pattern}{Logical, if TRUE treat files as regex patterns (default FALSE)}
}
\value{
An effectcheck object containing only results from specified files
}
\description{
Filters effectcheck results to show only results from specific files.
}
\examples{
results <- check_text("t(28) = 2.21, p = .035, d = 0.80")
filter_by_source(results, "text_input")
}
