% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{filter_by_test_type}
\alias{filter_by_test_type}
\title{Filter results by test type}
\usage{
filter_by_test_type(x, types)
}
\arguments{
\item{x}{An effectcheck object}

\item{types}{Character vector of test types to include (e.g., "t", "F", "r", "chisq", "z")}
}
\value{
An effectcheck object containing only the specified test types
}
\description{
Filters effectcheck results to show only specific test types.
}
\examples{
results <- check_text("t(28) = 2.21, p = .035. F(1, 50) = 4.03, p = .049")
filter_by_test_type(results, "t")
}
