% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_errors}
\alias{get_errors}
\title{Get errors from effectcheck results}
\usage{
get_errors(x)
}
\arguments{
\item{x}{An effectcheck object}
}
\value{
An effectcheck object containing only errors
}
\description{
Convenience function to extract only ERROR status results.
}
\examples{
results <- check_text("t(28) = 2.21, p = .035, d = 0.80")
get_errors(results)
}
