% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_warnings}
\alias{get_warnings}
\title{Get warnings from effectcheck results}
\usage{
get_warnings(x)
}
\arguments{
\item{x}{An effectcheck object}
}
\value{
An effectcheck object containing only warnings
}
\description{
Convenience function to extract only WARN status results.
}
\examples{
results <- check_text("t(28) = 2.21, p = .035, d = 0.80")
get_warnings(results)
}
