% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectcheck-class.R
\name{is.effectcheck}
\alias{is.effectcheck}
\title{Test if object is an effectcheck object}
\usage{
is.effectcheck(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Test if object is an effectcheck object
}
