% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectcheck-class.R
\name{rbind.effectcheck}
\alias{rbind.effectcheck}
\title{Combine effectcheck objects}
\usage{
\method{rbind}{effectcheck}(...)
}
\arguments{
\item{...}{effectcheck objects to combine}
}
\value{
Combined effectcheck object
}
\description{
Combine effectcheck objects
}
